requirements.txt
streamlit
text2vec
faiss-cpu
pandas
openpyxl